// src/app/Models/ipackage.ts
export interface IPackage {
  _id: string;
  name: string;
  arName: string;
  details: string;
  arDetails: string;
  price: number;
  numOfAds: number;
  discount: number;
  expiryDays: number;
  TitleadsDaysDetails: string;
  adsDaysDetails: string;
  image: string;
  arTitleadsDaysDetails:string;
  aradsDaysDetails:string;
}